preprocess/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
ca-certificates=2023.08.22=h06a4308_0
cudatoolkit=10.1.243=h6bb024c_0
freetype=2.12.1=h4a9f257_0
ftfy=6.1.3=pypi_0
giflib=5.2.1=h5eee18b_3
intel-openmp=2023.1.0=hdb19cb5_46306
joblib=1.3.2=pypi_0
jpeg=9b=h024ee3a_2
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h85742a9_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.4=h6a678d5_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
ncurses=6.4=h6a678d5_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openssl=3.0.12=h7f8727e_0
pillow=9.3.0=py38hace64e9_1
pip=23.3.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
pytorch=1.7.1=py3.8_cuda10.1.243_cudnn7.6.3_0
readline=8.2=h5eee18b_0
regex=2023.10.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.7.2=py38
torchvision=0.8.2=py38_cu101
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
wcwidth=0.2.12=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
